// verumoverview/backend/src/models/Person.ts
export interface Person {
  id_pessoa?: string;
  nome_completo: string;
  email: string;
  cargo_funcao?: string;
  time?: string;
  status?: string;
  perfil_comportamental?: any;
  engajamento?: number;
  projetos_vinculados?: any;
  anexos?: any;
  historico_movimentacoes?: any;
  comentarios?: any;
}